Revert "Refactor spec fixtures just a little"
